Allow vectors and ranges to compare and hash equal
AbstractArrays and ranges are now equal when their sizes and all their
elements are equal. In practice, this only applies to vectors since ranges
are one-dimensional.

hash() now also returns the same value for vectors and ranges. This is done
by comparing the first elements of vectors to the elements of a range with
the same length and first and last elements. All elements matching this
range at the beginning of the vector are hashed as a separate range,
and any remaining elements are hashed using a run-length encoding as before.